requirements.txt
aiohttp
aiosignal
anyio
asgiref
beautifulsoup4
certifi
charset-normalizer
click
crispy-bootstrap5
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-email
django-environ
django-extensions
django-filter
djangorestframework
djangorestframework_simplejwt
filelock
Flask
httpx
idna
Jinja2
joblib
MarkupSafe
numpy
opencv-contrib-python
packaging
pillow
psycopg2-binary
pydantic
PyJWT
qrcode
requests
scikit-learn
scipy
setuptools
six
soupsieve
sqlparse
supabase
threadpoolctl
typing_extensions
tzdata
urllib3
websockets
Werkzeug
